requirements.txt
pandas
numpy
scikit-learn
google-cloud-storage
gcsfs
joblib